OPTIMIZE test suite by refactoring into 'Feature' & 'Unit' directories
ADD use of sfneal/mock-models dev dependency.
CUT test models, factories & migrations that were replaced by sfneal/mock-models objects.
CUT migration & model factory tests.
