+++ composer.json
@@ -29,7 +29,8 @@
         "josiasmontag/laravel-redis-mock": ">=1.2.6",
         "orchestra/testbench": ">=6.7",
         "phpunit/phpunit": ">=7.5.20",
-        "scrutinizer/ocular": "^1.8"
+        "scrutinizer/ocular": "^1.8",
+        "sfneal/mock-models": ">=0.6"
     },
     "autoload": {
         "psr-4": {
