--- tests/ModelFactoryTest.php
@@ -1,55 +1,0 @@
-<?php
-
-namespace Sfneal\Datum\Tests;
-
-use Sfneal\Datum\Tests\Models\People;
-
-class ModelFactoryTest extends TestCase
-{
-    /**
-     * @var People
-     */
-    public $model;
-
-    /**
-     * Setup the test environment.
-     *
-     * @return void
-     */
-    public function setUp(): void
-    {
-        parent::setUp();
-        $this->model = People::query()->get()->shuffle()->first();
-    }
-
-    /** @test */
-    public function fillables_are_correct_types()
-    {
-        $this->assertIsString($this->model->name_first);
-        $this->assertIsString($this->model->name_last);
-        $this->assertStringContainsString('@', $this->model->email);
-        $this->assertIsInt($this->model->age);
-        $this->assertIsString($this->model->address);
-        $this->assertIsString($this->model->city);
-        $this->assertIsString($this->model->state);
-        $this->assertIsString($this->model->zip);
-    }
-
-    /** @test */
-    public function attributes_are_correct_types()
-    {
-        // Name attributes
-        $this->assertIsString($this->model->name_full);
-        $this->assertStringContainsString(', ', $this->model->name_last_first);
-        $this->assertStringContainsString($this->model->name_first, $this->model->name_full);
-        $this->assertStringContainsString($this->model->name_last, $this->model->name_full);
-
-        // Address attributes
-        $this->assertIsString($this->model->address_full);
-        $this->assertStringContainsString(', ', $this->model->address_full);
-        $this->assertStringContainsString($this->model->address, $this->model->address_full);
-        $this->assertStringContainsString($this->model->city, $this->model->address_full);
-        $this->assertStringContainsString($this->model->state, $this->model->address_full);
-        $this->assertStringContainsString($this->model->zip, $this->model->address_full);
-    }
-}
