--- tests/Queries/Traits/PeopleBuilder.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Sfneal\Datum\Tests\Queries\Traits;
-
-use Illuminate\Database\Eloquent\Builder;
-use Sfneal\Datum\Tests\Models\People;
-
-trait PeopleBuilder
-{
-    /**
-     * Retrieve a Query builder.
-     *
-     * @return Builder
-     */
-    protected function builder(): Builder
-    {
-        return People::query();
-    }
-}
